refactor: move resource allocator into cost estimator
