+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/scheduling/CostEstimator.scala
@@ -24,10 +24,13 @@ import edu.uci.ics.amber.core.tuple.Tuple
 import edu.uci.ics.amber.core.virtualidentity.ActorVirtualIdentity
 import edu.uci.ics.amber.core.workflow.WorkflowContext
 import edu.uci.ics.amber.engine.architecture.scheduling.DefaultCostEstimator.DEFAULT_OPERATOR_COST
+import edu.uci.ics.amber.engine.architecture.scheduling.SchedulingUtils.replaceVertex
+import edu.uci.ics.amber.engine.architecture.scheduling.resourcePolicies.ResourceAllocator
 import edu.uci.ics.amber.engine.common.AmberLogging
 import edu.uci.ics.texera.dao.SqlServer
 import edu.uci.ics.texera.dao.SqlServer.withTransaction
 import edu.uci.ics.texera.dao.jooq.generated.Tables.{WORKFLOW_EXECUTIONS, WORKFLOW_VERSION}
+import org.jgrapht.graph.DirectedAcyclicGraph
 
 import java.net.URI
 import scala.util.{Failure, Success, Try}
@@ -36,7 +39,15 @@ import scala.util.{Failure, Success, Try}
   * A cost estimator should estimate a cost of running a region under the given resource constraints as units.
   */
 trait CostEstimator {
-  def estimate(region: Region, resourceUnits: Int): Double
+
+  /**
+    * Uses the given resource units to allocate resources to the region, and determine a cost based on the allocation.
+    *
+    * Note currently the ResourceAllocator is not cost-based and thus we use a cost model that does not rely on the
+    * allocator, i.e., the cost estimation process is external to the ResourceAllocator.
+    * @return An updated region with allocated resources and an estimated cost of this region.
+    */
+  def allocateResourcesAndEstimateCost(region: Region, resourceUnits: Int): (Region, Double)
 }
 
 object DefaultCostEstimator {
@@ -50,6 +61,7 @@ object DefaultCostEstimator {
   */
 class DefaultCostEstimator(
     workflowContext: WorkflowContext,
+    val resourceAllocator: ResourceAllocator,
     val actorId: ActorVirtualIdentity
 ) extends CostEstimator
     with AmberLogging {
@@ -73,14 +85,21 @@ class DefaultCostEstimator(
     case Some(_) =>
   }
 
-  override def estimate(region: Region, resourceUnits: Int): Double = {
-    this.operatorEstimatedTimeOption match {
+  override def allocateResourcesAndEstimateCost(
+      region: Region,
+      resourceUnits: Int
+  ): (Region, Double) = {
+    // Currently the dummy cost from resourceAllocator is discarded.
+    val (newRegion, _) = resourceAllocator.allocate(region)
+    // We use a cost model that does not rely on the resource allocation.
+    // TODO: Once the ResourceAllocator actually calculates a cost, we can use its calculated cost.
+    val cost = this.operatorEstimatedTimeOption match {
       case Some(operatorEstimatedTime) =>
         // Use past statistics (wall-clock runtime). We use the execution time of the longest-running
         // operator in each region to represent the region's execution time, and use the sum of all the regions'
         // execution time as the wall-clock runtime of the workflow.
         // This assumes a schedule is a total-order of the regions.
-        val opExecutionTimes = region.getOperators.map(op => {
+        val opExecutionTimes = newRegion.getOperators.map(op => {
           operatorEstimatedTime.getOrElse(op.id.logicalOpId.id, DEFAULT_OPERATOR_COST)
         })
         val longestRunningOpExecutionTime = opExecutionTimes.max
@@ -89,11 +108,12 @@ class DefaultCostEstimator(
         // Without past statistics (e.g., first execution), we use number of ports needing storage as the cost.
         // Each port needing storage has a portConfig.
         // This is independent of the schedule / resource allocator.
-        region.resourceConfig match {
+        newRegion.resourceConfig match {
           case Some(config) => config.portConfigs.size
           case None         => 0
         }
     }
+    (newRegion, cost)
   }
 
   /**
