+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/scheduling/SchedulingUtils.scala
@@ -1,0 +1,57 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package edu.uci.ics.amber.engine.architecture.scheduling
+
+import org.jgrapht.graph.DirectedAcyclicGraph
+
+import scala.jdk.CollectionConverters.CollectionHasAsScala
+
+object SchedulingUtils {
+
+  def replaceVertex(
+      graph: DirectedAcyclicGraph[Region, RegionLink],
+      oldVertex: Region,
+      newVertex: Region
+  ): Unit = {
+    if (oldVertex.equals(newVertex)) {
+      return
+    }
+    graph.addVertex(newVertex)
+    graph
+      .outgoingEdgesOf(oldVertex)
+      .asScala
+      .toList
+      .foreach(oldEdge => {
+        val dest = graph.getEdgeTarget(oldEdge)
+        graph.removeEdge(oldEdge)
+        graph.addEdge(newVertex, dest, RegionLink(newVertex.id, dest.id))
+      })
+    graph
+      .incomingEdgesOf(oldVertex)
+      .asScala
+      .toList
+      .foreach(oldEdge => {
+        val source = graph.getEdgeSource(oldEdge)
+        graph.removeEdge(oldEdge)
+        graph.addEdge(source, newVertex, RegionLink(source.id, newVertex.id))
+      })
+    graph.removeVertex(oldVertex)
+  }
+}
